% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolecter_xt.R
\name{biolect_xt}
\alias{biolect_xt}
\title{Processing biolecter XT ouptut data}
\usage{
biolect_xt(
  path_to_biolector_xlsx,
  working_dir,
  caliberation = "both",
  replic = TRUE,
  n_sheets = 8
)
}
\arguments{
\item{path_to_biolector_xlsx}{an absolute path to the output file.}

\item{working_dir}{is an absolute path to the working directory for the output files.}

\item{replic}{takes bolean values, TRUE (default; you have biological or technical replica) and FALSE (no replica).}

\item{n_sheets}{the number of sheets in the output.xlsx file. This is important for wrangling the full dataset.}

\item{calibration}{can take three values; TRUE (only shows calibrated values), FALSE (only shows raw data), and both (shows both calibrated and raw data).}
}
\description{
Processing biolecter XT ouptut data
}
